requirements.txt
aiohttp==3.6.2
apiron==4.2.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
multidict==4.5.2
oauthlib==3.1.0
PySocks==1.7.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
slackclient==2.3.1
soupsieve==1.9.5
tweepy==3.8.0
urllib3==1.25.6
yarl==1.3.0
